ref: changed weight type to int, implemented autoincrement id

+++ src/main/java/models/animals/Carnivore.java
@@ -12,7 +12,7 @@ public class Carnivore extends Animal {
                      int currentAge,
                      boolean isAlive,
                      int maxAge,
-                     double weight,
+                     int weight,
                      int reproductiveRate,
                      Habitat mainHabitat,
                      AnimalType animalType,
